Static implementation of stack
A stack is a logical concept that consists of a set of similar elements. The term is often used in programming and memory organization in computers. Programming stacks are based on the principle of last in first out (LIFO), a commonly used type of data abstract that consists of two major operations, push and pop.